[Gold IV] Title: 좋다, Time: 132 ms, Memory: 12776 KB -BaekjoonHub

+++ 백준/Gold/1253. 좋다/좋다.java
@@ -24,7 +24,7 @@ public class Main {
             int right = n - 1;
 
             while(left < right){
-                long sum = arr[left] + arr[right];
+                int sum = arr[left] + arr[right];
 
                 if(left == i){
                     left++;
